requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_openai
langchain_google_genai
langchain_community
pypdf
faiss-cpu
structlog
PyMuPDF
structlog
pandas








-e . 